feat!: replace read/write with Base
I was encountering a problem where FITSFiles.jl was exporting
FITSFiles.write, which conflicts with Base.write. That doesn't seem to
be necessary, as the docs for Base.write (and Base.read) are both
compatible with the signatures used in FITSFiles.

Relevant documentation:

    https://docs.julialang.org/en/v1/base/io-network/#Base.write

    https://docs.julialang.org/en/v1/base/io-network/#Base.read

This commit changes all definitions of read and write to be Base.read
and Base.write respectively. Since all of these functions have type
signatures specific to the FITSFiles package (and correctly specify ::IO
as the first parameter), this ought not to suffer from any type piracy
or excessive compile invalidations.